+++ Tests/StreamChatUITests/SnapshotTests/Composer/ComposerVC_Tests.swift
@@ -733,6 +733,37 @@ final class ComposerVC_Tests: XCTestCase {
         }
     }
 
+    func test_didChangeLinks_whenEnrichSuccess_whenUrlDoesNotEqualToCurrentInput_thenDoNotCallShowPreview() {
+        let composerVC = SpyComposerVC()
+        composerVC.components.isComposerLinkPreviewEnabled = true
+        let mock = ChatChannelController_Mock.mock(client: .mock())
+        mock.channel_mock = .mockNonDMChannel(config: .mock(urlEnrichmentEnabled: true))
+        let mockAPIClient = mock.client.mockAPIClient
+        composerVC.channelController = mock
+        composerVC.enrichUrlDebouncer = .init(0, queue: .main)
+        composerVC.content = .initial()
+        composerVC.content.text = """
+        Some link: https://github.com/GetStream/stream-chat-swiftui
+        Another one: www.google.com
+        """
+        composerVC.updateContent()
+
+        let url = URL(string: "https://github.com/GetStream/stream-chat-swift")!
+        mockAPIClient.test_mockResponseResult(.success(LinkAttachmentPayload(
+            originalURL: url
+        )))
+
+        composerVC.didChangeLinks([
+            .init(url: url, originalText: "https://github.com/GetStream/stream-chat-swift", range: .init(location: 0, length: 0)),
+            .init(url: URL(string: "http://www.google.com")!, originalText: "www.google.com", range: .init(location: 0, length: 0))
+        ])
+
+        AssertAsync {
+            Assert.willBeEqual(composerVC.showLinkPreviewCallCount, 0)
+            Assert.willBeEqual(composerVC.dismissLinkPreviewCallCount, 0)
+        }
+    }
+
     func test_didChangeLinks_whenEnrichFails_thenDismissLinkPreview() {
         let composerVC = SpyComposerVC()
         composerVC.components.isComposerLinkPreviewEnabled = true
@@ -758,6 +789,35 @@ final class ComposerVC_Tests: XCTestCase {
 
         AssertAsync {
             Assert.willBeEqual(composerVC.dismissLinkPreviewCallCount, 1)
+        }
+    }
+
+    func test_didChangeLinks_whenEnrichFails_whenUrlDoesNotEqualToCurrentInput_thenDoNotCallDismissPreview() {
+        let composerVC = SpyComposerVC()
+        composerVC.components.isComposerLinkPreviewEnabled = true
+        let mock = ChatChannelController_Mock.mock(client: .mock())
+        mock.channel_mock = .mockNonDMChannel(config: .mock(urlEnrichmentEnabled: true))
+        let mockAPIClient = mock.client.mockAPIClient
+        composerVC.channelController = mock
+        composerVC.enrichUrlDebouncer = .init(0, queue: .main)
+        composerVC.content = .initial()
+        composerVC.content.text = """
+        Some link: https://github.com/GetStream/stream-chat-swiftui
+        Another one: www.google.com
+        """
+        composerVC.updateContent()
+
+        let url = URL(string: "https://github.com/GetStream/stream-chat-swift")!
+        mockAPIClient.test_mockResponseResult(Result<LinkAttachmentPayload, Error>.failure(ClientError()))
+
+        composerVC.didChangeLinks([
+            .init(url: url, originalText: "https://github.com/GetStream/stream-chat-swift", range: .init(location: 0, length: 0)),
+            .init(url: URL(string: "http://www.google.com")!, originalText: "www.google.com", range: .init(location: 0, length: 0))
+        ])
+
+        AssertAsync {
+            Assert.willBeEqual(composerVC.showLinkPreviewCallCount, 0)
+            Assert.willBeEqual(composerVC.dismissLinkPreviewCallCount, 0)
         }
     }